refact: 개발환경을 next dev 로 변경, 개발용 server 분리, deploy.js 추가
