+++ package.json
@@ -61,9 +61,10 @@
     "webpackbar": "^5.0.2"
   },
   "scripts": {
-    "dev": "cross-env NODE_ENV=development nodemon ./server.js",
+    "dev": "next dev",
+    "server": "cross-env NODE_ENV=development nodemon ./server.js",
     "build": "next build",
-    "deploy": "cross-env NODE_ENV=production node ./server.js",
+    "deploy": "cross-env NODE_ENV=production node ./deploy.js",
     "lint": "next lint",
     "sitemap": "next-sitemap --config ./next-sitemap.config.js"
   },
