+++ server.js
@@ -1,38 +1,14 @@
 const express = require("express");
-const next = require("next");
+const cors = require("cors");
 const jsonServer = require("json-server");
 const expressGraphQL = require("express-graphql");
 const schema = require("./schema/schema.js");
 
-const isDev = process.env.NODE_ENV === "development";
+const server = express();
+const port = process.env.PORT || 8080;
 
-const app = next({ dev: isDev });
+server.use(cors());
+server.use("/api", jsonServer.router("./db/db.json"));
+server.use("/graphql", expressGraphQL.graphqlHTTP({ graphiql: true, schema }));
 
-const handle = app.getRequestHandler();
-
-app.prepare().then(() => {
-  const server = express();
-
-  const port = process.env.PORT || 8080;
-
-  server.use("/sitemap.xml", function (_req, res) {
-    res.sendFile(__dirname + "/public/sitemap.xml");
-  });
-  server.use("/api", jsonServer.router("./db/db.json"));
-  server.use(
-    "/graphql",
-    expressGraphQL.graphqlHTTP({ graphiql: isDev, schema })
-  );
-
-  if (!isDev) {
-    server.get("/admin/**", (_req, res) => {
-      return res.redirect("/");
-    });
-  }
-
-  server.get("*", (req, res) => {
-    return handle(req, res);
-  });
-
-  server.listen(port, () => console.log(`Listening on port ${port}`));
-});
+server.listen(port, () => console.log(`Listening on port ${port}`));
